requirements.txt
cryptography
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-tinymce
easy-thumbnails
humanize
mysqlclient
Pillow
psycopg2-binary
pybase64
python-dateutil
python-decouple
python-dotenv
pytz
requests
sendgrid
six
sqlparse
twilio
weasyprint